moved components into separate folders and updated their routes
