+++ client/routes.js
@@ -4,9 +4,9 @@ import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
-import AllProducts from './components/AllProducts'
-import SingleProduct from './components/SingleProduct'
-import Cart from './components/Cart'
+import AllProducts from './components/Products/AllProducts'
+import SingleProduct from './components/Products/SingleProduct'
+import Cart from './components/Cart/Cart'
 
 /**
  * COMPONENT
